+++ CHANGELOG.md
@@ -3,6 +3,9 @@ This log follows the conventions of
 [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
+Nothing yet.
+
+## [Version 0.4.0] - 2020-11-29
 ### Changed
 - The default typeface for legends is now DejaVu Sans Mono. This should
   present excellent backwards compatibility with the previous default,
@@ -61,7 +64,8 @@ This log follows the conventions of
 ### Added
 - Minimal and maquette styles.
 
-[Unreleased]: https://github.com/veikman/dmote-keycap/compare/v0.3.0...HEAD
+[Unreleased]: https://github.com/veikman/dmote-keycap/compare/v0.4.0...HEAD
+[Version 0.4.0]: https://github.com/veikman/dmote-keycap/compare/v0.3.0...v0.4.0
 [Version 0.3.0]: https://github.com/veikman/dmote-keycap/compare/v0.2.0...v0.3.0
 [Version 0.2.0]: https://github.com/veikman/dmote-keycap/compare/v0.1.1...v0.2.0
 [Version 0.1.1]: https://github.com/veikman/dmote-keycap/compare/v0.1.0...v0.1.1